+++ source/search_engine_forum_TREC.h
@@ -8,5 +8,5 @@ private:
 	char run_id[MAX_RUN_NAME_LENGTH];
 public:
 	ANT_search_engine_forum_TREC(char *filename, char *participant_id, char *run_id, char *task);
-	void write(long topic_id, char **docids, long hits);
+	void write(long topic_id, char **docids, long long hits);
 } ;
